LPS-197489 Move classe FieldPredicateProviderTracker to class DefaultFilterFactoryImpl, and apply to usages
