--- modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/odata/filter/expression/field/predicate/provider/FieldPredicateProviderTracker.java
@@ -1,36 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.object.rest.internal.odata.filter.expression.field.predicate.provider;
-
-import com.liferay.object.odata.filter.expression.field.predicate.provider.FieldPredicateProvider;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Alejandro Tardín
- */
-@Component(service = FieldPredicateProviderTracker.class)
-public class FieldPredicateProviderTracker {
-
-	public FieldPredicateProvider getFieldPredicateProvider(String key) {
-		return _serviceTrackerMap.getService(key);
-	}
-
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		_serviceTrackerMap = ServiceTrackerMapFactory.openSingleValueMap(
-			bundleContext, FieldPredicateProvider.class,
-			"field.predicate.provider.key");
-	}
-
-	private ServiceTrackerMap<String, FieldPredicateProvider>
-		_serviceTrackerMap;
-
-}
